requirements.txt
omegaconf
Pillow
einops
pytorch-lightning
torcheval
lightning
opencv-python
pyclipper
shapely